+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Login/password authentication
 - Add profile page
 - Add verification email
+- Add password reset
 
 ### Changed